move routes to separate file and rename plugin => server
